Seq2SeqAttention/requirements.txt
Click==7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
matplotlib==3.1.1
numpy==1.16.3
pandas==0.24.2
Pillow==6.2.1
scikit-learn==0.21.3
scipy==1.3.1
tensorboard==2.0.1
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0

